Move https support with mkcert to a preset
