+++ packages/webpack-config/src/presets/index.js
@@ -1,7 +1,8 @@
 export { default as eslint } from './eslint.js';
+export { default as https } from './https.js';
 export { default as prototyping } from './prototyping.js';
 export { default as stylelint } from './stylelint.js';
 export { default as tailwindcss } from './tailwindcss.js';
 export { default as twig } from './twig.js';
-export { default as yaml } from './yaml.js';
 export { default as withContentHash } from './withContentHash.js';
+export { default as yaml } from './yaml.js';
